Add explanation for old copyright statement in MIT
At the time of writing the SPDX templates (at least for MIT) do not have
an option to treat different copyright statement formats as equal:

- "Copyright (c)" – in MIT, but not in line with US Copyright Office
- "Copyright" – in line with US Copyright Office
- "©" – in line with US Copyright Office

So in order to avoid confusion why it differs, an explanation was added.

An alternative solution would be to change it in the MIT example as well and
request SPDX to change the templates. This would probably be best in the long
run.