Make requirements apply to the same set of files
Before this change, the License File requirement applied to all files in
a Project, but the Copyright and Licensing Information requirement
applied to only some of the files in a Project. This change makes it so
that:

- If a file needs a License File, then it also needs Copyright and
  Licensing information.
- If a file needs Copyright and Licensing Information, then it also
  needs a License File.

Some files don't need to have either. This change also makes it so that
if you have one, then you need have the other.

Fixes #84.