update changelog with released changes after 3.0 release
As noted in #120, there are some changes that were introduced in the `stable`
branch (currently 3.0) after the actual release.

We consider all of them minor but it would have been good practice to make a
proper release. We will make sure this doesn't happen in the future. This
commit tracks the changes that occurred.